Struct quicr::Endpoint

pub struct Endpoint(_);

A QUIC endpoint.

An endpoint corresponds to a single UDP socket, may host many connections, and may act as both client and server for different connections.

Methods

impl Endpoint

pub fn new<'a>() -> EndpointBuilder<'a>

Begin constructing an Endpoint

Important traits for Box<R>
impl<R> Read for Box<R> where
R: Read + ?Sized, impl<W> Write for Box<W> where
W: Write + ?Sized, impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn connect(
&self,
addr: &SocketAddr,
hostname: Option<&[u8]>
) -> Box<Future<Item = (Connection, IncomingStreams), Error = ConnectionError>>

Connect to a remote endpoint.

hostname is used by the remote endpoint for disambiguation if addr hosts multiple services.