Trait quicr::Write

pub trait Write {
    fn poll_write(&mut self, buf: &[u8]) -> Poll<usize, WriteError>;

    fn poll_finish(&mut self) -> Poll<(), ConnectionError>;

    fn reset(&mut self, error_code: u16);
}

Trait of writable streams

Required Methods

fn poll_write(&mut self, buf: &[u8]) -> Poll<usize, WriteError>

Write some bytes to the stream.

fn poll_finish(&mut self) -> Poll<(), ConnectionError>

Indicate that no more data will be written.

Completes when the peer has acknowledged all sent data.

fn reset(&mut self, error_code: u16)

Abandon transmitting data on this stream.

No new data may be transmitted, and no previously transmitted data will be retransmitted if lost.