Crate quicr

QUIC transport protocol support for Tokio

QUIC is a modern transport protocol addressing shortcomings of TCP, such as head-of-line blocking, poor security, slow handshakes, and inefficient congestion control. This crate provides a portable userspace implementation.

The entry point of this crate is the Endpoint.

The futures and streams defined in this crate are not Send because they necessarily share state with eachother. As a result, they must be spawned on a single-threaded tokio runtime.

extern crate tokio;
extern crate quicr;
extern crate futures;

use futures::Future;

fn main() {
  let mut runtime = tokio::runtime::current_thread::Runtime::new().unwrap();
  let mut builder = quicr::Endpoint::new();
  // <configure builder>
  let (endpoint, driver, _) = builder.bind("[::]:0").unwrap();
  runtime.spawn(driver.map_err(|e| panic!("IO error: {}", e)));
  // ...
}

Structs

Config
Parameters governing the core QUIC state machine.

Connection
A QUIC connection.

ConnectionId
Protocol-level identifier for a connection.

Driver
A future that drives IO on an endpoint.

Endpoint
A QUIC endpoint.

EndpointBuilder
A helper for constructing an Endpoint.

IncomingStreams
A stream of QUIC streams initiated by a remote peer.

ListenKeys
Information that should be preserved between restarts for server endpoints.

NewConnection
A connection initiated by a remote client.

ReadToEnd
Future produced by read_to_end

RecvStream
A stream that can only be used to receive data

SendStream
A stream that can only be used to send data

Stream
A stream that supports both sending and receiving data

Enums

ConnectionError
Reasons why a connection might be lost.

Error
Errors that can occur during the construction of an Endpoint.

NewStream
A stream initiated by a remote peer.

ReadError
Errors that arise from reading from a stream.

WriteError
Errors that arise from writing to a stream

Traits

Read
Trait of readable streams

Write
Trait of writable streams

Functions

read_to_end
Uses unordered reads to be more efficient than using AsyncRead would allow

Type Definitions

Incoming
Stream of incoming connections.