Struct quicr::Connection

pub struct Connection(_);

A QUIC connection.

If a Connection is dropped without being explicitly closed, it will be automatically closed with an error_code of 0 and an empty reason.

Methods

impl Connection

Important traits for Box<R>
impl<R> Read for Box<R> where
R: Read + ?Sized, impl<W> Write for Box<W> where
W: Write + ?Sized, impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn open_uni(
&self
) -> Box<Future<Item = SendStream, Error = ConnectionError>>

Initite a new outgoing unidirectional stream.

Important traits for Box<R>
impl<R> Read for Box<R> where
R: Read + ?Sized, impl<W> Write for Box<W> where
W: Write + ?Sized, impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn open_bi(&self) -> Box<Future<Item = Stream, Error = ConnectionError>>

Initiate a new outgoing bidirectional stream.

Important traits for Box<R>
impl<R> Read for Box<R> where
R: Read + ?Sized, impl<W> Write for Box<W> where
W: Write + ?Sized, impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn close(
self,
error_code: u16,
reason: &[u8]
) -> Box<Future<Item = (), Error = ()>>

Close the connection immediately.

This does not ensure delivery of outstanding data. It is the application's responsibility to call this only when all important communications have been completed.

error_code and reason are not interpreted, and are provided directly to the peer.

reason will be truncated to fit in a single packet with overhead; to be certain it is preserved in full, it should be kept under 1KiB.

pub fn remote_address(&self) -> SocketAddr

The peer's UDP address.

pub fn local_id(&self) -> ConnectionId

The ConnectionId used for conn locally.

pub fn remote_id(&self) -> ConnectionId

The ConnectionId used for conn by the peer.

pub fn protocol(&self) -> Option<Box<[u8]>>

The negotiated application protocol